+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/FillSpell.java
@@ -26,7 +26,8 @@ public class FillSpell extends BrushSpell
 		{
 			return SpellResult.NO_TARGET;
 		}
-		if (!hasBuildPermission(targetBlock)) {
+        boolean hasPermission = brush.isErase() ? hasBreakPermission(targetBlock) : hasBuildPermission(targetBlock);
+        if (!hasPermission) {
 			return SpellResult.INSUFFICIENT_PERMISSION;
 		}