+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/GrenadeSpell.java
@@ -30,7 +30,7 @@ public class GrenadeSpell extends BlockSpell
 		if (target == null) {
 			return SpellResult.NO_TARGET;
 		}
-		if (!hasBuildPermission(target)) {
+		if (!hasBreakPermission(target)) {
 			return SpellResult.INSUFFICIENT_PERMISSION;
 		}